Moved getSpannedTitle from separate module to convert-title module. #28
