+++ client/utils/convert-title.js
@@ -1,3 +1,18 @@
+function getSpannedTitle() {
+  let $title = document.getElementById('main-title')
+  let $spannedTitle = document.createElement('h1')
+  $spannedTitle.setAttribute('id', 'animated-title')
+  let titleText = $title.textContent
+  for (let i = 0; i < titleText.length; i++) {
+    let $spanLetter = document.createElement('span')
+    $spanLetter.setAttribute('id', 'span-' + i)
+    $spanLetter.textContent = titleText[i]
+    $spanLetter.classList.add('span-title')
+    $spannedTitle.appendChild($spanLetter)
+  }
+  return $spannedTitle
+}
+
 function replaceTitleWithSpanned($spannedTitle) {
   let $title = document.getElementById('main-title')
   let $titleContainer = document.getElementById('title-container')
@@ -22,4 +37,4 @@ function reloadTitle() {
   $titleContainer.appendChild($title)
 }
 
-module.exports = { replaceTitleWithSpanned, replaceSpannedWithTitle, reloadTitle }
+module.exports = { getSpannedTitle, replaceTitleWithSpanned, replaceSpannedWithTitle, reloadTitle }
